+++ app/controllers/v1/base_controller.rb
@@ -1,0 +1,14 @@
+# rubocop:disable all
+class V1::BaseController < ApplicationController
+# rubocop:enable all
+  def api_update
+    @article ||= Article.find(params[:id])
+    @article.title = params[:post_title]
+    @article.text = params[:text]
+    if @article.save
+      render json: {}, status: 200
+    else
+      render json: {}, status: 400
+    end
+  end
+end
