+++ app/views/articles/index.html.slim
@@ -14,7 +14,7 @@ div.col-md-10
 			th colspan="3"
 		- @articles.each do |a|
 			tr
-				= form_for a, url: "/api/v1/articles/"+a.id.to_s+"/edit", remote:true, method: "post", f_class:'form', f_id: "id_"+a.id.to_s do |f|
+				= form_for a, remote:true, method: "post", f_class:'form', f_id: "id_"+a.id.to_s do |f|
 					td
 						= f.label :title, a.title, class:'lbl-title', id: "id_"+a.id.to_s
 						= f.text_field :post_title, class:'txt-title', style:'display:none', value: a.title, id: "id_"+a.id.to_s
