+++ public/docs/v1/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <title>Documentation for v1</title>
+    <link href="v1/style.css" media="screen" rel="stylesheet" type="text/css">
+  </head>
+  <body>
+    <div id="container">
+      <div id="operations">
+        <h3>API Operations</h3>
+      </div>
+      <div id="content">
+        <h1>Documentation for v1</h1>
+      </div>
+    </div>
+  </body>
+</html>
