+++ public/docs/v1/style.css
@@ -1,0 +1,5 @@
+body {margin: 0; background-color: #fff; color: #000; font-family: Arial,sans-serif;}
+content {margin-left: 200px;}
+content h1 {text-align: center;}
+operations {float: left; width: 200px; border-right: 1px solid #ccc;}
+operations h3 {text-align: center;}
