+++ spec/presenters/v1/base_presenter_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+describe V1::BasePresenter do
+end
