+++ spec/requests/v1/articles_controller_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+describe V1::ArticlesController do
+end
